+++ packages/bluesnap-direct-integration/src/bluesnap-direct-credit-card/bluesnap-direct-hosted-input-validator.spec.ts
@@ -1,9 +1,10 @@
 import { CREDIT_CARD_ERRORS } from '../bluesnap-direct-constants';
-import BlueSnapHostedInputValidator from './bluesnap-direct-hosted-input-validator';
 import {
     BlueSnapDirectHostedFieldTagId as HostedFieldTagId,
     BlueSnapDirectErrorDescription as SubmitErrorDescription,
 } from '../types';
+
+import BlueSnapHostedInputValidator from './bluesnap-direct-hosted-input-validator';
 
 describe('BlueSnapHostedInputValidator', () => {
     let validator: BlueSnapHostedInputValidator;